Fixed search bar
Didn't realize project was nested in open-recipes folder and had some issues with dependencies. working searchbar component example in index.tsx
