+++ open-recipes/src/app/(tabs)/index.tsx
@@ -1,4 +1,4 @@
-import { View,Image, StyleSheet, Platform } from 'react-native';
+import { View, Image, StyleSheet, Platform } from 'react-native';
 
 import { HelloWave } from '@/src/components/HelloWave';
 import ParallaxScrollView from '@/src/components/ParallaxScrollView';
@@ -7,34 +7,18 @@ import { ThemedView } from '@/src/components/ThemedView';
 import SearchBar from '@/src/components/SearchBar';
 
 export default function HomeScreen() {
-  const handleSearch = (query: string) => {
-    console.log('Searching for:', query);
-  };
 
-  const handleMenuPress = () => {
-    console.log('Menu pressed');
-    
-  };
 
   return (
-
     <ParallaxScrollView
       headerBackgroundColor={{ light: '#A1CEDC', dark: '#1D3D47' }}
       headerImage={
         <Image
-          source={require('@/assets/images/partial-react-logo.png')}
+          source={require('@/assets/images/partial-react-logo.png')} // Fixed: Added require
           style={styles.reactLogo}
         />
       }>
-            <View style={styles.container}>
-    <SearchBar 
-      onSearch={handleSearch}
-      onMenuPress={handleMenuPress}
-    />
-    {
-      
-    }
-  </View>
+   
       <ThemedView style={styles.titleContainer}>
         <ThemedText type="title">Welcome!</ThemedText>
         <HelloWave />
@@ -86,9 +70,9 @@ const styles = StyleSheet.create({
     bottom: 0,
     left: 0,
     position: 'absolute',
-  },  
+  },
   container: {
     flex: 1,
     backgroundColor: '#fff',
-  }
+  },
 });
